Clip Spawns
There's a spot on the table that the clips spawn in, and when u pick one up one will spawn in. Maybe will rework so instead of a designated spot on the table it's on ur side that u can grab. Eventually going to make it that if a clip hits the ground itll get destoryed. Also am disabling ray interaction as thats the issue with the double grab but I have no luck in figuring out how to disable it