State machine for spawner
The spawner will no longer be disabled and will start in an idle/pre-wave state and hitting the button should switch to a wave/spawning state. Then when all enemies are killed it should switch to a end-wave state in which the upgrade menu should appear. When the upgrade menu closes it should switch to the pre-wave state. Finally hitting the button again should be the loop.